Improve device and platform selection
Each predador-prey version now only needs to specify the type of
device to use, without specifying the platform. The platform and
device are automatically chosen given a type of device. In case
more than one device matching the specified type is found, the
program asks the user to select which device to use.