Fix bug in queue release, improve file utils and predpreycommon
There was a bug when releasing queues because zone.numQueues never
got initialized.
File utils uses fread instead of getchar in while, function is
renamed loadSource (instead of importKernel) and is nicer.
In PredPreyCommon, the printing of program log (when failed) now
uses dynamically allocated array instead of static array.