[OCLC] Newly born agents not "seen" by wolves actions
Previously newly born sheep were seen and could be eaten by wolves,
although due to a bug when adding newly born agents to the current
cell's agent list, this didn't had an impact on simulation
accuracy.

Fix is implemented by adding newly born agents to a separate list,
and only when the loop over agents in cell is over, is the newly
born agent list merged with the list of agents in the cell.